cópia-dos-repositórios/HackatonLLM/requirements.txt
Flask
google-generativeai
python-dotenv
gunicorn